tegra: nvmap: make sciipc_id as 64 bit
Make SciIpc_Id as 64 bit.

Bug 3715617

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2783704
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/nvmap/nvmap_sci_ipc.h
@@ -3,7 +3,7 @@
  *
  * mapping between nvmap_hnadle and sci_ipc entery
  *
- * Copyright (c) 2019-2021, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2019-2022, NVIDIA CORPORATION. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms and conditions of the GNU General Public License,
@@ -25,13 +25,13 @@ int nvmap_validate_sci_ipc_params(struct nvmap_client *client,
 int nvmap_create_sci_ipc_id(struct nvmap_client *client,
 				struct nvmap_handle *h,
 				u32 flags,
-				u32 *sci_ipc_id,
+				u64 *sci_ipc_id,
 				NvSciIpcEndpointVuid pr_vuid,
 				bool is_ro);
 
 int nvmap_get_handle_from_sci_ipc_id(struct nvmap_client *client,
 				u32 flags,
-				u32 sci_ipc_id,
+				u64 sci_ipc_id,
 				NvSciIpcEndpointVuid localusr_vuid,
 				u32 *h);
 #endif /*  __VIDEO_TEGRA_NVMAP_SCI_IPC_H */
